backend/requirements.txt
# requirements.txt

# --- Flask Core and Web Services ---
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
psycopg2-binary
python-dotenv
requests

# --- AI Services ---
openai-whisper
rasa